PyTest DQ Framework/requirements.txt
pytest~=8.3.3
pytest-html~=4.1.1
psycopg2-binary~=2.9.10
pandas~=2.2.3
pyarrow~=19.0.1
sqlalchemy~=2.0.36
python-dotenv~=1.0.1
